feat: add filter menus and clear filters button to AgendaView
