+++ iosApp/iosApp/Agenda/AgendaView.swift
@@ -116,122 +116,8 @@ struct AgendaView: View {
 
             .navigationBarTitle(L10n.screenAgenda)
             .navigationBarItems(trailing:
-                                    Menu("\(Image(systemName: "line.3.horizontal.decrease.circle"))") {
-                // Multi-value filter menus using SessionFilter
-                Menu("\(Image(systemName: "line.3.horizontal.decrease.circle"))") {
-                    // Favorites filter (single select)
-                    Button(action: {
-                        var newFilters = viewModel.sessionFilters
-                        let filter = SessionFilter(type: .bookmark, value: "bookmark")
-                        if newFilters.contains(filter) {
-                            newFilters.remove(filter)
-                        } else {
-                            newFilters.insert(filter)
-                        }
-                        viewModel.sessionFilters = newFilters
-                    }) {
-                        Label(L10n.filterFavorites, systemImage: viewModel.sessionFilters.contains { $0.type == .bookmark } ? "star.fill" : "star")
-                    }
-                    // Language filter (multi-select)
-                    Menu(L10n.sessionFiltersDrawerLanguagesLabel) {
-                        ForEach([SessionLanguage.french.rawValue, SessionLanguage.english.rawValue], id: \ .self) { lang in
-                            Button(action: {
-                                var newFilters = viewModel.sessionFilters
-                                let filter = SessionFilter(type: .language, value: lang)
-                                if newFilters.contains(filter) {
-                                    newFilters.remove(filter)
-                                } else {
-                                    newFilters.insert(filter)
-                                }
-                                viewModel.sessionFilters = newFilters
-                            }) {
-                                HStack {
-                                    Image(systemName: viewModel.sessionFilters.contains { $0.type == .language && $0.value == lang } ? "checkmark.square" : "square")
-                                    Text(lang == SessionLanguage.french.rawValue ? L10n.languageFrench : L10n.languageEnglish)
-                                }
-                            }
-                        }
-                    }
-                    // Complexity filter (multi-select)
-                    Menu(L10n.sessionFiltersDrawerComplexityLabel) {
-                        ForEach([Complexity.beginner.rawValue, Complexity.intermediate.rawValue, Complexity.advanced.rawValue], id: \ .self) { comp in
-                            Button(action: {
-                                var newFilters = viewModel.sessionFilters
-                                let filter = SessionFilter(type: .complexity, value: comp)
-                                if newFilters.contains(filter) {
-                                    newFilters.remove(filter)
-                                } else {
-                                    newFilters.insert(filter)
-                                }
-                                viewModel.sessionFilters = newFilters
-                            }) {
-                                HStack {
-                                    Image(systemName: viewModel.sessionFilters.contains { $0.type == .complexity && $0.value == comp } ? "checkmark.square" : "square")
-                                    Text(
-                                        comp == Complexity.beginner.rawValue ? L10n.complexityBeginer :
-                                        comp == Complexity.intermediate.rawValue ? L10n.complexityIntermediate :
-                                        L10n.complexityAdvanced
-                                    )
-                                }
-                            }
-                        }
-                    }
-                    // Room filter (multi-select)
-                    if let rooms = viewModel.roomsContent {
-                        Menu(L10n.sessionFiltersDrawerRoomsLabel) {
-                            ForEach(rooms, id: \ .id) { room in
-                                Button(action: {
-                                    var newFilters = viewModel.sessionFilters
-                                    let filter = SessionFilter(type: .room, value: room.id)
-                                    if newFilters.contains(filter) {
-                                        newFilters.remove(filter)
-                                    } else {
-                                        newFilters.insert(filter)
-                                    }
-                                    viewModel.sessionFilters = newFilters
-                                }) {
-                                    HStack {
-                                        Image(systemName: viewModel.sessionFilters.contains { $0.type == .room && $0.value == room.id } ? "checkmark.square" : "square")
-                                        Text(room.name)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    // Session type filter (multi-select)
-                    Menu(L10n.sessionFiltersDrawerTypeLabel) {
-                        ForEach([SessionType.conference.rawValue, SessionType.quickie.rawValue, SessionType.codelab.rawValue], id: \ .self) { type in
-                            Button(action: {
-                                var newFilters = viewModel.sessionFilters
-                                let filter = SessionFilter(type: .type, value: type)
-                                if newFilters.contains(filter) {
-                                    newFilters.remove(filter)
-                                } else {
-                                    newFilters.insert(filter)
-                                }
-                                viewModel.sessionFilters = newFilters
-                            }) {
-                                HStack {
-                                    Image(systemName: viewModel.sessionFilters.contains { $0.type == .type && $0.value == type } ? "checkmark.square" : "square")
-                                    Text(
-                                        type == SessionType.conference.rawValue ? L10n.sessionTypeConference :
-                                        type == SessionType.quickie.rawValue ? L10n.sessionTypeQuickie :
-                                        L10n.sessionTypeCodelab
-                                    )
-                                }
-                            }
-                        }
-                    }
-                    // Clear filters button
-                    if !viewModel.sessionFilters.isEmpty {
-                        Button(action: {
-                            viewModel.clearFilters()
-                        }) {
-                            Label(L10n.filterClear, systemImage: "trash")
-                        }
-                    }
-                }                
-            })
+                AgendaNavigationBarItems(viewModel: viewModel)
+            )
             .task {
                 RCValues.sharedInstance.fetchCloudValues()
                 await viewModel.observeRooms()
